--- lib/css_highlight.js
@@ -1,5 +1,0 @@
-var hl = require('highlight.js');
-
-module.exports = function (css) {
-    return hl.highlight('css', css).value;
-}
